Break out TheoreticalDatabase code, so we can load in that data from save and load later.
